Split Recordset components into separate files
